pyproject.toml
[tool.poetry]
name = "camel-oasis"
version = "0.1.0"
description = "Open Agents Social Interaction Simulations on a Large Scale"
authors = ["CAMEL-AI.org"]
readme = "README.md"
license = "Apache License 2.0"
keywords = [
    "recommendation-system",
    "communicative-ai",
    "ai-societies",
    "social-simulations",
    "artificial-intelligence",
    "deep-learning",
    "multi-agent-systems",
    "natural-language-processing",
    "large-language-models",
]
packages = [
    { include = "oasis" },
]
repository = "https://github.com/rutschmanna/oasis.git"

[tool.poetry.dependencies]
python = ">=3.10.0,<3.12"
pandas = "2.2.2"
igraph = "0.11.6"
cairocffi = "1.7.1"
pytest = "8.2.0"
pytest-asyncio = "0.23.6"
pillow = "10.3.0"
unstructured = "0.13.7"
sentence-transformers = "3.0.0"
pre-commit = "3.7.1"
prance = "23.6.21.0"
openapi-spec-validator = "0.7.1"
slack_sdk = "3.31.0"
neo4j = "5.23.0"
camel-ai = "0.2.43"
requests_oauthlib = "2.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
